some detail fixes and readme

+++ scripts/create-db.sql
@@ -4,10 +4,9 @@ CREATE TABLE agenda (
   id bigint(11) NOT NULL,
   name varchar(255) NOT NULL,
   email varchar(255) NOT NULL,
-  phone int(13) NOT NULL,
+  phone bigint(13) NOT NULL,
   registered datetime NULL,
   modified datetime NULL
-
 );
 
 ALTER TABLE agenda
